requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
docker>=6.1.2
psutil>=5.9.0
torch>=2.0.0
torchserve>=0.8.1
scikit-learn>=1.0.2
joblib>=1.1.0
boto3>=1.28.0
google-cloud-compute>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.22.0
azure-mgmt-compute>=29.1.0
azure-identity>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0 